+++ CHANGELOG.md
@@ -127,6 +127,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * New `subject(pattern)` revset function that matches first line of commit
   descriptions.
 
+* Conditional configuration now supports `--when.command` to change configuration
+  based on subcommand.
+
 ### Fixed bugs
 
 * Fixed diff selection by external tools with `jj split`/`commit -i FILESETS`.